[draft] init progress of making ScalarType headeronly
[ghstack-poisoned]
