+++ torch/headeronly/util/Float8_fnuz_cvt.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include <c10/util/floating_point_utils.h>
+#include <torch/headeronly/util/floating_point_utils.h>
 
 #include <cstdint>
 
@@ -8,7 +8,7 @@
 #include <sycl/sycl.hpp>
 #endif
 
-namespace c10::detail {
+namespace torch::headeronly::detail {
 
 /*
  * Convert a 8-bit floating-point number in either f8 E4M3FNUZ or bf8 E5M2FNUZ
@@ -61,4 +61,4 @@ inline C10_HOST_DEVICE float fp8_fnuz_to_fp32_value(uint8_t x) {
   return fp32_from_bits(retval);
 }
 
-} // namespace c10::detail
+} // namespace torch::headeronly::detail
